experiments/fastapi-rag-coffee/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-dotenv==1.0.1
psycopg[binary]==3.2.1
psycopg2-binary==2.9.9
httpx==0.27.2
pydantic==2.9.2
